+++ tests/unit/stochastic/test_stochastic_rocket.py
@@ -1,0 +1,25 @@
+from rocketpy.rocket.rocket import Rocket
+
+
+def test_str(stochastic_calisto):
+    assert isinstance(str(stochastic_calisto), str)
+
+
+def test_create_object(stochastic_calisto):
+    """Test create object method of StochasticRocket class.
+
+    This test checks if the create_object method of the StochasticCalisto
+    class creates a StochasticCalisto object from the randomly generated
+    input arguments.
+
+    Parameters
+    ----------
+    stochastic_calisto : StochasticCalisto
+        StochasticCalisto object to be tested.
+
+    Returns
+    -------
+    None
+    """
+    obj = stochastic_calisto.create_object()
+    assert isinstance(obj, Rocket)